Upgraded to new pattern

+++ ui/lib/src/elements/event-proposal-detail.ts
@@ -2,9 +2,9 @@ import { localized, msg } from '@lit/localize';
 import { customElement } from 'lit/decorators.js';
 import { html } from 'lit';
 
-import '@holochain-open-dev/elements/elements/display-error.js';
-import '@holochain-open-dev/profiles/elements/agent-avatar.js';
-import '@holochain-open-dev/file-storage/elements/show-image.js';
+import '@holochain-open-dev/elements/dist/elements/display-error.js';
+import '@holochain-open-dev/profiles/dist/elements/agent-avatar.js';
+import '@holochain-open-dev/file-storage/dist/elements/show-image.js';
 import '@shoelace-style/shoelace/dist/components/card/card.js';
 import '@shoelace-style/shoelace/dist/components/alert/alert.js';
 import '@shoelace-style/shoelace/dist/components/icon-button/icon-button.js';
@@ -14,9 +14,9 @@ import '@shoelace-style/shoelace/dist/components/spinner/spinner.js';
 import '@shoelace-style/shoelace/dist/components/skeleton/skeleton.js';
 
 import { consume } from '@lit-labs/context';
-import '@darksoil/assemble/elements/call-to-action-detail.js';
+import '@darksoil/assemble/dist/elements/call-to-action-detail.js';
 //@ts-ignore
-import { CallToActionDetail } from '@darksoil/assemble/elements/call-to-action-detail.js';
+import { CallToActionDetail } from '@darksoil/assemble/dist/elements/call-to-action-detail.js';
 import { wrapPathInSvg } from '@holochain-open-dev/elements';
 import { toPromise } from '@holochain-open-dev/stores';
 import { mdiAlarm, mdiCalendarClock, mdiCash, mdiMapMarker } from '@mdi/js';
@@ -27,32 +27,31 @@ import { ActionHash } from '@holochain/client';
 import { GatherStore } from '../gather-store';
 import { gatherStoreContext } from '../context';
 
-//@ts-ignore
 @localized()
-//@ts-ignore
 @customElement('event-proposal-detail')
 export class EventProposalDetail extends CallToActionDetail {
   /**
    * @internal
    */
-  //@ts-ignore
   @consume({ context: gatherStoreContext, subscribe: true })
   gatherStore!: GatherStore;
 
   async createCollectiveCommitment(satisfactions_hashes: Array<ActionHash>) {
     await super.createCollectiveCommitment(satisfactions_hashes);
 
-    const callToAction: EntryRecord<CallToAction> = await toPromise(
-      //@ts-ignore
+    const callToAction: EntryRecord<CallToAction> | undefined = await toPromise(
       this.assembleStore.callToActions.get(this.callToActionHash)
     );
+
+    if (!callToAction) return;
+
     const customContent = decode(callToAction.entry.custom_content) as any;
 
     await this.gatherStore.client.createEvent({
       title: callToAction.entry.title,
       ...customContent,
     });
-    //@ts-ignore
+
     // TODO: improve this so that we don't need to delete the calltoaction
     await this.assembleStore.client.deleteCallToAction(this.callToActionHash);
   }
@@ -73,7 +72,7 @@ export class EventProposalDetail extends CallToActionDetail {
         <div style="display: flex; flex-direction: row;">
           <div class="column" style="justify-content: end">
             <div
-              title=${msg("location")}
+              title=${msg('location')}
               style="display: flex; flex-direction: row; align-items: center;"
             >
               <sl-icon
@@ -86,7 +85,7 @@ export class EventProposalDetail extends CallToActionDetail {
             </div>
 
             <div
-              title=${msg("time")}
+              title=${msg('time')}
               style="display: flex; flex-direction: row; align-items: center"
             >
               <sl-icon
@@ -94,9 +93,8 @@ export class EventProposalDetail extends CallToActionDetail {
                 .src=${wrapPathInSvg(mdiCalendarClock)}
               ></sl-icon>
               <span style="white-space: pre-line"
-                >${new Date(
-                  customContent.start_time / 1000
-                ).toLocaleString()} - ${new Date(
+                >${new Date(customContent.start_time / 1000).toLocaleString()} -
+                ${new Date(
                   customContent.end_time / 1000
                 ).toLocaleString()}</span
               >
@@ -104,7 +102,7 @@ export class EventProposalDetail extends CallToActionDetail {
 
             ${customContent.cost
               ? html` <div
-                  title=${msg("cost")}
+                  title=${msg('cost')}
                   style="display: flex; flex-direction: row; align-items: center"
                 >
                   <sl-icon
@@ -115,20 +113,19 @@ export class EventProposalDetail extends CallToActionDetail {
                     >${customContent.cost}</span
                   >
                 </div>`
-              : html``
-            }
-
+              : html``}
             ${customContent.expiry_date
               ? html` <div
-                  title=${msg("expiry")}
+                  title=${msg('expiry')}
                   style="display: flex; flex-direction: row; align-items: center"
                 >
                   <span style="white-space: pre-line"
-                    ><b>${msg("Decision deadline:")}</b> ${new Date(customContent.expiry_date / 1000).toLocaleString()}</span
+                    ><b>${msg('Decision deadline:')}</b> ${new Date(
+                      customContent.expiry_date / 1000
+                    ).toLocaleString()}</span
                   >
                 </div>`
-              : html``
-            }
+              : html``}
           </div>
         </div>
       </div>
